requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Celery
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# Utilities
croniter==2.0.1
httpx==0.25.2
python-dateutil==2.8.2
django-environ==0.11.2
pytz==2023.3
